Fix cuDNN version regex.
The current regular expression is meant to strip the trailing '-1' from
the cuDNN version number, however, it also removes any occurrence of '1'
from the middle of the version number resulting in versions such as
'9.7.' (note the trailing dot).

This commit fixes the regex to print the version number leading up
to the trailing '-'.